LayoutTests/fast/js/kde/string-2-n-expected.txt

PASS window.successfullyParsed is undefined.
